+++ CHANGELOG.md
@@ -16,3 +16,4 @@ All notable changes to `strapi-wrapper` will be documented in this file.
 
 - Small fixes to catch meta responses not being set on strapi records
 - Add ability to set custom endpoint after the collection has been initialised
+- Fix an issue when trying to squash a non array dataset